Zmiana typu zwacacej wartości w metodzie Podziel na double

+++ sol1/pro1/Program.cs
@@ -5,7 +5,6 @@ Console.WriteLine("Hello, World!");
 Console.WriteLine(Dodaj(1,1));
 Console.WriteLine(CalculateAverage([1, 2, 3, 4, 5, 6, 7, 8]));
 
-
 int Dodaj(int liczba1, int liczba2)
 {
     return liczba1 + liczba2;
@@ -18,9 +17,9 @@ int Pomnoz(int liczba1, int liczba2)
 {
     return liczba1 * liczba2;
 }
-int Podziel(int liczba1, int liczba2)
+double Podziel(int liczba1, int liczba2)
 {
-    return liczba1 / liczba2;
+    return liczba1 / (double)liczba2;
 }
 double CalculateAverage(int[] values)
 {
